requirements.txt
Flask
gunicorn
firebase-admin
google-genai
moviepy